requirements.txt
discord.py>=2.2.2
requests~=2.28.2
SQLAlchemy~=2.0.7
psycopg2>=2.9.5
pycryptodome>=3.14